Refactored code and added a new file which contains summary generation logic
