+++ src/main.js
@@ -22,31 +22,10 @@ import asoc from './asoc.js';
 import settings from './settings.js';
 import statusChecker from './statusChecker.js';
 import resultProcessor from './resultProcessor.js';
+import summaryWriter from './summaryWriter.js';
 
 let sastScanId;
 let scaScanId;
-
-/*
- detect PR context
-*/
-const isPR = process.env.INPUT_IS_PR_SCAN === "true" || process.env.GITHUB_EVENT_NAME === "pull_request";
-/*
- PR metadata from workflow inputs*/
-const prNumber = process.env.INPUT_PR_NUMBER || "";
-const branchName = process.env.INPUT_BRANCH_NAME || process.env.GITHUB_HEAD_REF || process.env.GITHUB_REF_NAME || "";
-const repoName = process.env.INPUT_REPO_NAME || process.env.GITHUB_REPOSITORY || "";
-const commitSha = process.env.GITHUB_SHA || "";
-core.info(`GitHub Event: ${process.env.GITHUB_EVENT_NAME}`);
-if (isPR) {
-    core.info("Running SAST scan for Pull Request");
-    core.info(`Repository: ${repoName}`);
-    core.info(`PR Number: ${prNumber}`);
-    core.info(`Branch: ${branchName}`);
-    core.info(`Commit: ${commitSha}`);
-}
-else {
-    core.info("Running SAST scan for Push/Manual trigger");
-}
 
 core.info(constants.DOWNLOADING_CLIENT);
 saclientutil.downloadClient()
@@ -79,25 +58,24 @@ saclientutil.downloadClient()
         core.info(`SCA Scan ID: ${scaScanId}`)
         core.info(`${settings.getScanUrl(scaScanId)}`);
     }
-	
-	 /*
-         PR metadata logging
-    */
-	if(isPR) {
-            core.info(`PR scan detected`);
-            core.info(`PR branch: ${process.env.GITHUB_HEAD_REF}`);
-	}
     
     if(process.env.INPUT_WAIT_FOR_ANALYSIS !== 'true') {
-        return null;
+		const promises = [];
+		if(sastScanId) {
+			promises.push(summaryWriter.generateMinimumSummary(asoc.getScanDetails, sastScanId, "SAST"));
+		}
+		if(scaScanId) {
+			promises.push(summaryWriter.generateMinimumSummary(asoc.getScanDetails, scaScanId, "SCA"));
+		}
+		return Promise.all(promises).then(() => false);
     }
 
     core.info(constants.WAIT_FOR_ANALYSIS);
     return statusChecker.waitForAnalysis(sastScanId, scaScanId);
 })
 .then((timedOut) => {
-	if(timedOut === null) {
-		return null;
+	if(timedOut) {
+		return;
 	}
     if(timedOut) {
         core.warning(constants.ANALYSIS_TIMEOUT);
