+++ src/summaryWriter.js
@@ -1,0 +1,178 @@
+/*
+Copyright 2026 HCL America, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import * as fs from 'fs';
+import * as core from '@actions/core';
+import settings from './settings.js';
+
+function getGitHubContext() {
+    const isPR = process.env.GITHUB_EVENT_NAME === 'pull_request';
+    const repoName = process.env.GITHUB_REPOSITORY || "";
+    const branchName = process.env.GITHUB_HEAD_REF || process.env.GITHUB_REF_NAME || "";
+    const commitSha = process.env.GITHUB_SHA ? process.env.GITHUB_SHA.substring(0, 7) : "";
+    let prNumber = "";
+    try {
+        if (process.env.GITHUB_EVENT_PATH && fs.existsSync(process.env.GITHUB_EVENT_PATH)) {
+            const eventPayload = JSON.parse(fs.readFileSync(process.env.GITHUB_EVENT_PATH, "utf8"));
+            prNumber = eventPayload.pull_request?.number || "";
+        }
+    } catch (e) {
+        console.log("Failed to read PR information:", e.message);
+    }
+    return {isPR, repoName, branchName, commitSha, prNumber};
+}
+
+function writeSummaryMarkdown(summaryData, reportDownloadLink) {
+	const {total, counts, scanId, scanUrl, appName, appUrl, scanTime, scanType, isPR, repoName, branchName, commitSha, prNumber} = summaryData;
+	const scanLabel = isPR ? `${scanType} PR Scan Summary` : `${scanType} Scan Summary`;
+	const prUrl = `https://github.com/${repoName}/pull/${prNumber}`;
+	const branchUrl = `https://github.com/${repoName}/tree/${branchName}`;
+	const commitUrl = `https://github.com/${repoName}/commit/${process.env.GITHUB_SHA}`;
+	const prSection = isPR ? `
+## Pull Request Information
+
+| Field | Value |
+|------|------|
+| PR Number | [#${prNumber}](${prUrl}) |
+| Branch | [${branchName}](${branchUrl}) |
+| Commit | [${commitSha}](${commitUrl}) |
+
+---`
+: "";
+    const enableHyperlinks = process.env.INPUT_SCAN_INFO_HYPERLINKS !== "false";
+	const scanIdValue = enableHyperlinks ? `[${scanId}](${scanUrl})` : scanId;
+	const appNameValue = enableHyperlinks ? `[${appName}](${appUrl})` : appName;
+	const reportLabel = `Download ${scanType} HTML Report`;
+	const reportValue = enableHyperlinks ? `[${reportLabel}](${reportDownloadLink})` : "";
+	const md = `
+# HCL AppScan ${scanLabel}
+
+${prSection}
+
+### Scan Information
+
+| Field | Value |
+|------|-------|
+| Scan Type | ${scanType} |
+| Scan ID | ${scanIdValue} |
+| Application Name | ${appNameValue} |
+| Repository | ${repoName} |
+| Scan Time | ${scanTime} |
+
+---
+
+## Total Vulnerabilities: ${total}
+
+| Critical | High | Medium | Low | Info |
+|----------|------|--------|-----|------|
+| ${counts.Critical} | ${counts.High} | ${counts.Medium} | ${counts.Low} | ${counts.Informational} |
+
+---
+
+${reportValue}
+
+`;
+     const mdFileName = isPR ? `appscan-${scanType.toLowerCase()}-pr-report.md` : `appscan-${scanType.toLowerCase()}-build-summary-report.md`;
+	 fs.writeFileSync(mdFileName, md);
+	 if (process.env.GITHUB_STEP_SUMMARY) {
+		 fs.appendFileSync(process.env.GITHUB_STEP_SUMMARY, md);
+	}
+}
+
+async function generateMinimumSummary(getScanDetails, scanId, scanType) {
+    const summaryContext = await getSummaryContext(getScanDetails, scanId, scanType);
+	const {scanUrl, appUrl, appName, scanTime, isPR, repoName, branchName, commitSha, prNumber} = summaryContext;
+    const enableHyperlinks = process.env.INPUT_SCAN_INFO_HYPERLINKS !== "false";
+    const scanIdValue = enableHyperlinks ? `[${scanId}](${scanUrl})` : scanId;
+    const appNameValue = enableHyperlinks ? `[${appName}](${appUrl})` : appName;
+	const prUrl = `https://github.com/${repoName}/pull/${prNumber}`;
+	const branchUrl = `https://github.com/${repoName}/tree/${branchName}`;
+	const commitUrl = `https://github.com/${repoName}/commit/${process.env.GITHUB_SHA}`;
+	const prSection = isPR ? `
+## Pull Request Information
+
+| Field | Value |
+|------|------|
+| PR Number | [#${prNumber}](${prUrl}) |
+| Branch | [${branchName}](${branchUrl}) |
+| Commit | [${commitSha}](${commitUrl}) |
+
+---`
+: "";	
+    const md = `
+# HCL AppScan ${scanType} ${isPR ? "PR " : ""}Scan Summary
+
+${prSection}
+
+### Scan Information
+
+| Field | Value |
+|--------|-------|
+| Scan Type | ${scanType} |
+| Scan ID | ${scanIdValue} |
+| Application Name | ${appNameValue} |
+| Repository | ${repoName} |
+| Scan Time | ${scanTime} |
+
+---
+
+`;
+
+    const mdFileName = `appscan-${scanType.toLowerCase()}-minimum-summary.md`;
+    fs.writeFileSync(mdFileName, md);
+    if (process.env.GITHUB_STEP_SUMMARY) {
+        fs.appendFileSync(process.env.GITHUB_STEP_SUMMARY, md);
+    }
+}
+
+async function buildIssueSummary(getScanDetails, groupedItems, scanId, scanType) {
+	const counts = {Critical: 0, High: 0, Medium: 0, Low: 0, Informational: 0};
+	groupedItems.forEach(item => {
+		if (counts[item.Severity] !== undefined) {
+			counts[item.Severity] += Number(item.N || 0);
+		}
+    });
+	const total = Object.values(counts).reduce((a,b)=>a+b, 0);
+	const summaryContext = await getSummaryContext(getScanDetails, scanId, scanType);
+	return({total, counts, ...summaryContext});
+}
+
+async function getSummaryContext(getScanDetails, scanId, scanType) {
+    const serviceUrl = settings.getServiceUrl();
+    const baseUrl = serviceUrl.replace("/api/v4", "");
+    const applicationId = process.env.INPUT_APPLICATION_ID;
+    const scanUrl = `${baseUrl}/main/myapps/${applicationId}/scans/${scanId}`;
+    const appUrl = `${baseUrl}/main/myapps/${applicationId}`;
+    let appName = applicationId;
+    try {
+        let scanDetails = null;
+        if (scanType === "SAST") {
+            scanDetails = await getScanDetails(scanId, "Sast");
+        } else if (scanType === "SCA") {
+            scanDetails = await getScanDetails(scanId, "Sca");
+        }
+        if (scanDetails) {
+            appName = scanDetails.AppName || appName;
+        }
+    }
+    catch (e) {
+        console.log("Failed to fetch AppName from scan details:", e.message);
+    }
+    const scanTime = new Date().toISOString().replace("T", " ").substring(0, 19);
+    return {scanId, scanUrl, appName, appUrl, scanTime, scanType, ...getGitHubContext()};
+}
+
+export default { generateMinimumSummary, writeSummaryMarkdown, getGitHubContext, buildIssueSummary }
